+++ packages/asc-ui/src/components/AmsterdamLogo/AmsterdamLogoStyle.ts
@@ -1,7 +1,7 @@
 import LogoShort from '@datapunt/asc-assets/static/icons/LogoShort.svg'
 import LogoTall from '@datapunt/asc-assets/static/icons/LogoTall.svg'
 import styled, { css } from '@datapunt/asc-core'
-import { breakpoint, focusStyleOutline } from '../../utils'
+import { breakpoint, focusStyleOutline, themeSpacing } from '../../utils'
 
 export type Props = {
   tall?: boolean
@@ -19,7 +19,8 @@ const tallStyle = css`
   @media screen and ${breakpoint('min-width', 'laptopM')} {
     height: 68px;
     width: 100px;
-
+    margin-right: ${themeSpacing(10)};
+    
     ${LogoStyle} {
       background-image: url("${LogoTall}");
     }
@@ -31,6 +32,7 @@ const AmsterdamLogoStyle = styled.a<Props>`
   height: 30px;
   width: 68px;
   flex-shrink: 0;
+  margin-right: ${themeSpacing(3)};
   ${focusStyleOutline()}
   ${({ tall }) => tall && tallStyle}
 `